temporally restructuring for gemspec
