--- helpers/helper.rb
@@ -1,5 +1,0 @@
-helpers do
-  def app_root
-    "#{env['rack.url_scheme']}://#{env['HTTP_HOST']}#{env['SCRIPT_NAME']}"
-  end
-end
