+++ lib/slackattend/update_by_slack.rb
