+++ lib/slackattend/version.rb
@@ -1,0 +1,3 @@
+module Slackattend
+  VERSION = "0.0.1"
+end
