+++ lib/views/index.haml
