+++ lib/views/layout.haml
